Create Unsafe class for future deletion
